+++ test/Compze.Internals.Logging.Specifications/When_logging_to_the_SerilogLogger_backend.cs
@@ -23,7 +23,7 @@ public class When_logging_to_the_SerilogLogger_backend
    protected When_logging_to_the_SerilogLogger_backend()
    {
       var serilog = new LoggerConfiguration().MinimumLevel.Verbose().WriteTo.Sink(Sink).CreateLogger();
-      Logger1 = SerilogLogger.Create(GetType(), serilog).WithLogLevel(LogLevel.Debug);
+      Logger1 = SerilogLogger.Create(GetType(), serilog).WithLogLevel(LogLevel.Trace);
    }
 
    public class an_interpolated_message_with_named_holes : When_logging_to_the_SerilogLogger_backend
@@ -60,22 +60,24 @@ public class When_logging_to_the_SerilogLogger_backend
    {
       public an_interpolated_message_at_each_level()
       {
+         var t = 0; Logger1.Trace($"t {t}");
          var x = 1; Logger1.Debug($"d {x}");
          var y = 2; Logger1.Info($"i {y}");
          var z = 3; Logger1.Warning($"w {z}");
          var q = 4; Logger1.Warning(new InvalidOperationException("warn-ex"), $"we {q}");
          var r = 5; Logger1.Error(new InvalidOperationException("err-ex"), $"e {r}");
+         var c = 6; Logger1.Critical(new InvalidOperationException("crit-ex"), $"c {c}");
       }
 
-      [XF] public void each_level_arrives_at_Serilog_with_its_own_level()
-         => Sink.Events.Select(it => it.Level).Must().SequenceEqual([LogEventLevel.Debug, LogEventLevel.Information, LogEventLevel.Warning, LogEventLevel.Warning, LogEventLevel.Error]);
+      [XF] public void each_level_arrives_at_Serilog_with_its_own_level_Compze_Trace_becoming_Verbose_and_Compze_Critical_becoming_Fatal()
+         => Sink.Events.Select(it => it.Level).Must().SequenceEqual([LogEventLevel.Verbose, LogEventLevel.Debug, LogEventLevel.Information, LogEventLevel.Warning, LogEventLevel.Warning, LogEventLevel.Error, LogEventLevel.Fatal]);
 
       [XF] public void each_template_carries_the_original_named_holes()
          => Sink.Events.Select(it => it.MessageTemplate.Text)
-                       .Must().SequenceEqual(["d {x}", "i {y}", "w {z}", "we {q}", "e {r}"]);
+                       .Must().SequenceEqual(["t {t}", "d {x}", "i {y}", "w {z}", "we {q}", "e {r}", "c {c}"]);
 
-      [XF] public void exceptions_passed_to_Warning_and_Error_are_attached_to_the_corresponding_log_events()
+      [XF] public void exceptions_passed_to_Warning_Error_and_Critical_are_attached_to_the_corresponding_log_events()
          => Sink.Events.Select(it => it.Exception?.Message)
-                       .Must().SequenceEqual([null, null, null, "warn-ex", "err-ex"]);
+                       .Must().SequenceEqual([null, null, null, null, "warn-ex", "err-ex", "crit-ex"]);
    }
 }